Ensure project_eval result is received by the parent (#45)
Currently the child does `process.send` and immediately `process.exit`.
My understanding is that `process.send` is async and it is not
guaranteed that the message is sent by the time the child exists. To
guarantee it, we make the parent send an acknowledgement before exiting.
See https://github.com/nodejs/node/issues/6767.