Support TanStack Start (#49)
TanStack Start uses Vite, so technically it already works. However, logs
from server functions and API routes are not captured, since they run in
an isolated runtime, separate from the Vite server where we serve the
MCP. This PR changes the logger from in-memory to file based, so that it
works in the multi-runtime scenario.

This also adds an entrypoint specific to for TanStack, which is
necessary to enable log capturing. I added a setup section to the
README.